feat(#844): remove publish workflow from game settings
